Added StackedMap data structure
A stack of Map's that itself implements a Map interface.  Its
function as a Map is of a flattened version of itself, containing
all keys/values from all maps in the stack.  Where key collisions
occur between maps in the stack, the value closer to the top of
the stack wins.