Split Configuration interface from simple implementation
Apps that use this lib may need to do something more complex from
a config perspective.  Maybe they want to return a new instance
of filter/decorator/publisher each time they are asked for.  Maybe
a filter is not thread safe and they want to create a new filter
for each thread asking for configuration.  Some other scenarios
as well.